# sem11/openapi.yaml
openapi: 3.0.3
info:
  title: Robot service 3.0
  description: |-
    API сервис управления роботом - пылесосом

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: https://localhost:9669
...
paths:
...
  /Robot:
    put:
      tags:
        - Robot
      summary: Обновить робота
      description: Обновить робота
      operationId: updateRobot
      requestBody:
        description: Создание нового робота в БД
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Robot'
          application/xml:
            schema:
              $ref: '#/components/schemas/Robot'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Robot'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Robot'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Robot'
        '400':
          description: Invalid ID supplied
        '404':
          description: Robot not found
        '405':
          description: Validation exception
          
    post:
      tags:
        - Robot
      summary: Добавить робота
      description: Добавить робота БД
      operationId: addRobot
      requestBody:
        description: Создание нового робота в БД
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Robot'
          application/xml:
            schema:
              $ref: '#/components/schemas/Robot'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Robot'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Robot'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Robot'
        '405':
          description: Invalid input
  /Robot/findById/{robotId}:
    get:
      tags:
        - Robot
      summary: Найти робота по ID
      description: Вернуть одного робота по ID
      operationId: getRobotById
      parameters:
        - name: robotId
          in: path
          description: Идентификатор робота
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Robot'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Robot'
        '400':
          description: Invalid ID supplied
        '404':
          description: Robot not found
  /Robot/deleteById/{robotId}:      
    delete:
      tags:
        - Robot
      summary: Удаляет робота
      description: Удаляет робота
      operationId: deleteRobot
      parameters:
        - name: robotId
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: robotId
          in: path
          description: ID робота
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid Robot value
  /Schedule:
    put:
      tags:
        - Schedule
      summary: Обновить расписание
      description: Обновить расписание по ID
      operationId: updateSchedule
      requestBody:
        description: Обновить расписание в БД
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
          application/xml:
            schema:
              $ref: '#/components/schemas/Schedule'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Schedule'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          description: Invalid ID supplied
        '404':
          description: Schedule not found
        '405':
          description: Validation exception
          
    post:
      tags:
        - Schedule
      summary: Добавить расписание
      description: Добавить расписание
      operationId: addSchedule
      requestBody:
        description: Добавить расписание в БД
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Schedule'
          application/xml:
            schema:
              $ref: '#/components/schemas/Schedule'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Schedule'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'  
            application/xml:
              schema:
                $ref: '#/components/schemas/Schedule'
        '405':
          description: Invalid input
  /Schedule/findById/{scheduleId}:
    get:
      tags:
        - Schedule
      summary: Найти расписание по ID
      description: Вернуть расписание по ID
      operationId: getScheduleById
      parameters:
        - name: scheduleId
          in: path
          description: Идентификатор расписания
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Schedule'          
            application/xml:
              schema:
                $ref: '#/components/schemas/Schedule'
        '400':
          description: Invalid ID supplied
        '404':
          description: Schedule not found
  /Schedule/deleteById/{id}:      
    delete:
      tags:
        - Schedule
      summary: Удалить расписание
      description: Удалить расписание по ID
      operationId: deleteSchedule
      parameters:
        - name: id
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: id
          in: path
          description: ID расписания
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid Schedule value
...
components:
  schemas:
...
    Schedule:
      required:
        - dateTime
        - mode
        - robotId
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        dateTime:
          type: string
          example: dd.mm.yyyy-hh:mm:ss
        mode:
          type: integer
          format: int64
          example: 10
        robotId:
          type: integer
          format: int64
          example: 10
      xml:
        name: Schedule
...
    Robot:
      required:
        - name
        - model
        - versionPO
        - charge
        - garbageContainer
        - robotPollution
        - nextService
        - serialNumber
        - ipAddress
        - groupId
        - clianingId
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: something
        model:
          type: string
          example: something
        versionPO:
          type: string
          example: something
        charge:
          type: integer
          format: int64
          example: 10
        garbageContainer:
          type: integer
          format: int64
          example: 10
        robotPollution:
          type: integer
          format: int64
          example: 10
        nextService:
          type: integer
          format: int64
          example: 10
        serialNumber:
          type: string
          example: something
        ipAddress:
          type: integer
          format: int64
          example: 10
        groupId:
          type: integer
          format: int64
          example: 10
        clianingId:
          type: integer
          format: int64
          example: 10  
      xml:
        name: Robot
...