# sem11/src/spring-server-generated/src/main/java/io/swagger/api/RobotApiController.java
package io.swagger.api;

import io.swagger.model.Robot;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponses;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.bind.annotation.CookieValue;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.bind.annotation.RestController;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import javax.servlet.http.HttpServletRequest;
import java.io.IOException;
import java.util.List;
import java.util.Map;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2023-08-12T11:28:20.035575131Z[GMT]")
@RestController
public class RobotApiController implements RobotApi {

    private static final Logger log = LoggerFactory.getLogger(RobotApiController.class);

    private final ObjectMapper objectMapper;

    private final HttpServletRequest request;

    @org.springframework.beans.factory.annotation.Autowired
    public RobotApiController(ObjectMapper objectMapper, HttpServletRequest request) {
        this.objectMapper = objectMapper;
        this.request = request;
    }

    public ResponseEntity<Robot> addRobot(@Parameter(in = ParameterIn.DEFAULT, description = "Создание нового робота в БД", required=true, schema=@Schema()) @Valid @RequestBody Robot body) {
        String accept = request.getHeader("Accept");
        if (accept != null && accept.contains("application/json")) {
            try {
                return new ResponseEntity<Robot>(objectMapper.readValue("{\n  \"charge\" : 10,\n  \"serialNumber\" : \"something\",\n  \"nextService\" : 10,\n  \"versionPO\" : \"something\",\n  \"groupId\" : 10,\n  \"name\" : \"something\",\n  \"ipAddress\" : 10,\n  \"model\" : \"something\",\n  \"id\" : 10,\n  \"garbageContainer\" : 10,\n  \"clianingId\" : 10,\n  \"robotPollution\" : 10\n}", Robot.class), HttpStatus.NOT_IMPLEMENTED);
            } catch (IOException e) {
                log.error("Couldn't serialize response for content type application/json", e);
                return new ResponseEntity<Robot>(HttpStatus.INTERNAL_SERVER_ERROR);
            }
        }

        return new ResponseEntity<Robot>(HttpStatus.NOT_IMPLEMENTED);
    }

    public ResponseEntity<Void> deleteRobot(@Parameter(in = ParameterIn.PATH, description = "ID робота", required=true, schema=@Schema()) @PathVariable("robotId") Long robotId,@Parameter(in = ParameterIn.HEADER, description = "" ,schema=@Schema()) @RequestHeader(value="robotId", required=false) String robotId) {
        String accept = request.getHeader("Accept");
        return new ResponseEntity<Void>(HttpStatus.NOT_IMPLEMENTED);
    }

    public ResponseEntity<Robot> getRobotById(@Parameter(in = ParameterIn.PATH, description = "Идентификатор робота", required=true, schema=@Schema()) @PathVariable("robotId") Long robotId) {
        String accept = request.getHeader("Accept");
        if (accept != null && accept.contains("application/json")) {
            try {
                return new ResponseEntity<Robot>(objectMapper.readValue("{\n  \"charge\" : 10,\n  \"serialNumber\" : \"something\",\n  \"nextService\" : 10,\n  \"versionPO\" : \"something\",\n  \"groupId\" : 10,\n  \"name\" : \"something\",\n  \"ipAddress\" : 10,\n  \"model\" : \"something\",\n  \"id\" : 10,\n  \"garbageContainer\" : 10,\n  \"clianingId\" : 10,\n  \"robotPollution\" : 10\n}", Robot.class), HttpStatus.NOT_IMPLEMENTED);
            } catch (IOException e) {
                log.error("Couldn't serialize response for content type application/json", e);
                return new ResponseEntity<Robot>(HttpStatus.INTERNAL_SERVER_ERROR);
            }
        }

        return new ResponseEntity<Robot>(HttpStatus.NOT_IMPLEMENTED);
    }

    public ResponseEntity<Robot> updateRobot(@Parameter(in = ParameterIn.DEFAULT, description = "Создание нового робота в БД", required=true, schema=@Schema()) @Valid @RequestBody Robot body) {
        String accept = request.getHeader("Accept");
        if (accept != null && accept.contains("application/json")) {
            try {
                return new ResponseEntity<Robot>(objectMapper.readValue("{\n  \"charge\" : 10,\n  \"serialNumber\" : \"something\",\n  \"nextService\" : 10,\n  \"versionPO\" : \"something\",\n  \"groupId\" : 10,\n  \"name\" : \"something\",\n  \"ipAddress\" : 10,\n  \"model\" : \"something\",\n  \"id\" : 10,\n  \"garbageContainer\" : 10,\n  \"clianingId\" : 10,\n  \"robotPollution\" : 10\n}", Robot.class), HttpStatus.NOT_IMPLEMENTED);
            } catch (IOException e) {
                log.error("Couldn't serialize response for content type application/json", e);
                return new ResponseEntity<Robot>(HttpStatus.INTERNAL_SERVER_ERROR);
            }
        }

        return new ResponseEntity<Robot>(HttpStatus.NOT_IMPLEMENTED);
    }

}